requirements.txt
requests
spacy
python-dotenv
numpy
pandas
flask
streamlit
-e .

